Sentiment analysis (was fun working on)
Built on the RNN (Recurrent Neural Network) as the input is a sequential stream of data. 
Generated the Word Mappings from the scratch using Prediction based encoding.
Both the context as well as the semantics of the words in the data were retained.
Skip gram algorithm was used to generate these word mappings.
Tensorflow's PCA package was used to reduce the dimensionality of the data.
Used python powered tensorflow and Wor2Vec model was employed to get the prediction based word embedding done which internally functioned based on the skipgram model (used the target word to predict and generate the context words). 
Softmax Predictor was used as the last layer of interpreter for the output.
Further added Gradient descent function to reduce the error from the output of the hidden layers.
Data set used: IMDB reviews.